rename ValueObjectFactory to mapper
